backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-dotenv==1.0.0
playwright==1.50.0
psycopg2-binary==2.9.10
